requirements.txt
aiohttp==3.8.1
aiomultiprocess==0.9.0
aiosignal==1.2.0
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
Django==3.2
frozenlist==1.3.1
future==0.18.2
h11==0.12.0
httpcore==0.15.0
httplib2==0.20.4
httpx==0.23.0
idna==3.3
jieba==0.42.1
kafka-python==2.0.2
multidict==6.0.2
Naked==0.1.31
nebula3-python==3.1.0
numpy==1.23.1
pandas==1.4.3
pycrypto==2.6.1
pycryptodome==3.9.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
redis==2.10.6
requests==2.28.1
rfc3986==1.5.0
shellescape==3.8.1
six==1.12.0
sniffio==1.3.0
sqlparse==0.4.2
urllib3==1.26.12
yarl==1.8.1
